requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-openai
faiss-cpu
python-dotenv
pandas
tabulate
python-pptx
openpyxl
supabase